pyproject.toml
[tool.poetry]
name = "rosneft"
version = "0.1.0"
description = ""
authors = ["rodion"]

[tool.poetry.dependencies]
python = "^3.9"
Django = "^4.1.5"
djangorestframework = "^3.14.0"
graphene-django = "^3.0.0"
django-filter = "^22.1"
psycopg2-binary = "^2.9.5"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
